+++ src/workflow/consts.ts
@@ -1,0 +1,17 @@
+export const moveProjectWorkflow = [
+  { title: "Project Name", value: "projectName" },
+  { title: "Project Type", value: "projectType" },
+  { title: "Network", value: "network" },
+];
+
+export const fullstackProjectWorkflow = [
+  { title: "Project Name", value: "projectName" },
+  { title: "Project Type", value: "projectType" },
+  { title: "Template", value: "template" },
+  { title: "Network", value: "network" },
+];
+
+export enum ProjectType {
+  MOVE = "move",
+  FULLSTACK = "fullstack",
+}